+++ src/main/java/za/ac/sun/cs/coastal/strategy/tracebased/RandomTestingFactory.java
@@ -251,7 +251,7 @@ public class RandomTestingFactory extends TraceBasedFactory {
 				return ThreadLocalRandom.current().nextLong(min, max + 1);
 			}
 		}
-		
+
 	}
 
 }